feat: adding documentation stories to Layout components
