+++ src/components/layout/Footer/Footer.stories.jsx
@@ -1,0 +1,12 @@
+import Footer from "./Footer";
+
+export default {
+  component: Footer,
+  title: 'Argatio/Layout/Footer',
+  tags: ['autodocs'],
+  layout: {
+    background: 'dark',
+  },
+};
+
+export const Default = {};
