+++ src/components/layout/Navbar/NavBar.stories.jsx
@@ -1,0 +1,16 @@
+import NavBar from "./NavBar";
+
+export default {
+  component: NavBar,
+  title: 'Argatio/Layout/NavBar',
+  tags: ['autodocs'],
+  decorators: [
+    (Story) => (
+      <div className="bg-black">
+        <Story />
+      </div>
+    )
+  ]
+};
+
+export const Default = {};
